+++ foxtive/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Foxtive Changelog
 Foxtive changelog file 
 
+### 0.20.0 (2025-01-15)
+* feat(http): add start_datetime & end_datetime to query params
+* bump(crates): to latest versions
+
 ### 0.19.5 (2025-10-31)
 * feat(global-state): add is_initialized method and document methods